refactor: move ArduinoSerialReceiver to repository package
